chore: run prettier on all files in pre-commit
Currently, we're only running prettier on Markdown, YAML, JSON, and CSS
files. We're purposely excluding JavaScript files, as that would cause a
race-condition with ESLint, which also runs on JavaScript files.

However, we can enable prettier on **ALL** files, as long as set the
`--concurrent false` option on `lint-staged`, in order to avoid any
race conditions.